# docs/swagger.yaml
info:
  contact:
    email: 2313773899@qq.com`
    name: 黄日明`
  description: 为了自己以后方便写点小东西，搭个模版`
  license:
    name: Apache 2.0`
    url: '[http://www.apache.org/licenses/LICENSE-2.0.html](http://www.apache.org/licenses/LICENSE-2.0.html)`'
  title: 项目模版`
  version: 1.0`
paths:
  /api/classify/query-classify:
    get:
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: classify_name
        in: query
        name: classify_name
        type: string
      - description: rank
        in: query
        name: rank
        type: string
      - description: is_use
        in: query
        name: is_use
        type: number
      produces:
      - application/json
      summary: 查询分类
  /api/classify/query-classify-type:
    get:
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: classify_name
        in: query
        name: classify_name
        type: string
      - description: rank
        in: query
        name: rank
        type: string
      - description: is_use
        in: query
        name: is_use
        type: number
      produces:
      - application/json
      summary: 查询分类以及对应二级分类
  /api/company/query-company:
    get:
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: company_name
        in: query
        name: company_name
        type: string
      - description: is_use
        in: query
        name: is_use
        type: number
      produces:
      - application/json
      summary: 查询公司
  /api/type/query-type:
    get:
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: classify_name
        in: query
        name: classify_name
        type: string
      - description: type_name
        in: query
        name: type_name
        type: string
      - description: classify_id
        in: query
        name: classify_id
        type: string
      - description: is_use
        in: query
        name: is_use
        type: number
      produces:
      - application/json
      summary: 查询二级分类
  /api/user-login/layout:
    get:
      summary: 登出
  /api/user-login/wxLogin:
    post:
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        type: string
      - description: 登陆来源
        in: query
        name: come_from
        type: string
      - description: 登陆方式
        in: query
        name: login_type
        type: string
      - description: 用户名
        in: query
        name: nick_name
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 邮箱
        in: query
        name: email
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      produces:
      - application/json
      summary: 登陆相关方法,分开两个方法是为了避免以后或许想对应不同端登陆可以做一些不同的操作
  /api/user-login/wxapp-dncrypt:
    post:
      parameters:
      - description: rawData
        in: query
        name: rawData
        type: string
      - description: sessionKey
        in: query
        name: sessionKey
        type: string
      - description: iv
        in: query
        name: iv
        type: string
      produces:
      - application/json
      summary: 微信解密
  /api/user-login/wxapp-get-openid:
    post:
      parameters:
      - description: openid
        in: query
        name: code
        type: string
      produces:
      - application/json
      summary: 微信获取openid和session_key,为了数据安全，不应该把session_key返回前端
  /api/user/account-manage/binding-phone:
    post:
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: email
        in: query
        name: email
        type: string
      - description: nick_name
        in: query
        name: nick_name
        type: string
      - description: come_from
        in: query
        name: come_from
        type: string
      - description: gender
        in: query
        name: gender
        type: string
      - description: city
        in: query
        items:
          type: string
        name: city
        type: array
      - description: province
        in: query
        items:
          type: string
        name: province
        type: array
      - description: country
        in: query
        items:
          type: string
        name: country
        type: array
      - description: language
        in: query
        items:
          type: string
        name: language
        type: array
      - description: is_admin
        in: query
        name: is_admin
        type: string
      - description: birthday
        in: query
        items:
          type: string
        name: birthday
        type: array
      - description: create_at
        in: query
        items:
          type: string
        name: create_at
        type: array
      - description: delete_at
        in: query
        items:
          type: string
        name: delete_at
        type: array
      - description: update_at
        in: query
        items:
          type: string
        name: update_at
        type: array
      - description: is_use
        in: query
        name: is_use
        type: string
      - description: pageNum
        in: query
        name: pageNum
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      summary: 查询用户信息
swagger: "2.0"